+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Changed
 - Fix black text on dark background when dark mode is used
 - Fix searching from multiple sources being performed serially
+- UI loading speed improvements
 
 ## [0.2.0] - 2021-08-08
 ### Added